feat(hackathon): support multiple events at /hackathon/<slug>
Each event is now its own page under src/pages/hackathon/, reusing the
shared HackathonEventPage template or a fully custom layout. The
/hackathon index redirects to the env-selected active event
(HACKATHON_EVENT_SLUG) and renders a placeholder when unset.

The announcement banner is now purely env-driven
(HACKATHON_BANNER_ENABLED, HACKATHON_EVENT_SLUG, HACKATHON_BANNER_TEXT);
the date-window logic (HACKATHON_START/HACKATHON_END) is removed.
