+++ CONTRIBUTING.md
@@ -38,22 +38,23 @@ SHOW_DRAFTS=true
 
 A flag is **enabled only when its value is exactly `"true"`** — any other value (empty, `"1"`, `"yes"`) is treated as disabled.
 
-### Hackathon Banner
+### Hackathon Banner & Events
 
-The site-wide announcement bar that points visitors at [`/hackathon`](./src/pages/hackathon.tsx) is driven by env vars at build time and resolved by `src/lib/hackathon-banner-server.ts`. The banner uses Docusaurus' built-in `themeConfig.announcementBar` so it shows above the navbar on every page (docs, templates, solutions, MDX). It's intentionally **non-dismissible** so the banner stays as the only on-site entry point to `/hackathon` for the full event window.
+Each hackathon event is its own page at `/hackathon/<slug>`, served by a file under [`src/pages/hackathon/`](./src/pages/hackathon/) (for example [`apps-agents-for-good-2026.tsx`](./src/pages/hackathon/apps-agents-for-good-2026.tsx)). Events are fully independent — editing one never touches another. An event can reuse the shared [`HackathonEventPage`](./src/components/hackathon/hackathon-event-page.tsx) template by passing a typed `HackathonEvent` object, or render a completely bespoke layout instead. Event pages are `noindex` and kept out of `sitemap.xml`; entry is via the banner.
 
-| Env var                    | Purpose                                                                                                                                                                                                                                                                 |
-| -------------------------- | ----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------- |
-| `HACKATHON_BANNER_ENABLED` | Escape hatch. `"true"` forces the banner on, `"false"` forces it off. Anything else (including unset) falls through to dates.                                                                                                                                           |
-| `HACKATHON_START`          | ISO date `YYYY-MM-DD`. Required (along with `HACKATHON_END`) for date-driven activation.                                                                                                                                                                                |
-| `HACKATHON_END`            | ISO date `YYYY-MM-DD`. Inclusive through 23:59:59.999 UTC of the given day.                                                                                                                                                                                             |
-| `HACKATHON_BANNER_TEXT`    | Optional override for the **lead-in text only** (HTML allowed). The "See resources" link to `/hackathon` is always appended, so a misconfigured override can never strand visitors on a banner with no way in. Defaults to `"Databricks Developer Hackathon is live."`. |
+`/hackathon` ([`src/pages/hackathon/index.tsx`](./src/pages/hackathon/index.tsx)) redirects to the active event named by `HACKATHON_EVENT_SLUG`. When no slug is set it shows a minimal placeholder instead of redirecting nowhere.
 
-Precedence: `HACKATHON_BANNER_ENABLED` wins. Otherwise the banner is active if both dates are set, parseable, `start <= end`, and "now" is within the window.
+The site-wide announcement bar is driven by env vars at build time and resolved by [`src/lib/hackathon-banner-server.ts`](./src/lib/hackathon-banner-server.ts). It uses Docusaurus' built-in `themeConfig.announcementBar` so it shows above the navbar on every page (docs, templates, solutions, MDX). It's intentionally **non-dismissible** so it stays the only on-site entry point to the event for the full window.
 
-The `/hackathon` page itself is always live (and discoverable in `sitemap.xml`) so the URL is shareable ahead of the event. It is not linked from the navbar or the footer — the banner is the only on-site entry point, and it only appears while the banner is active.
+| Env var                    | Purpose                                                                                                                                                                                                                                                 |
+| -------------------------- | ------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------- |
+| `HACKATHON_BANNER_ENABLED` | Banner on **only when the value is exactly `"true"`** — any other value (including unset) is off. Same convention as `SHOW_DRAFTS`.                                                                                                                     |
+| `HACKATHON_EVENT_SLUG`     | Slug of the active event. The banner links to `/hackathon/<slug>` and `/hackathon` redirects there. With no slug the banner links to `/hackathon` (which then shows the placeholder).                                                                   |
+| `HACKATHON_BANNER_TEXT`    | Optional override for the **lead-in text only** (HTML allowed). The "See resources" link is always appended, so a misconfigured override can never strand visitors on a banner with no way in. Defaults to `"Databricks Developer Hackathon is live."`. |
 
-Flipping the banner on Vercel is "edit env var → redeploy", the same model as `SHOW_DRAFTS`. Bump the `id` in `getHackathonBannerConfig` per event so prior dismissals (stored in `localStorage`) are reset for returning visitors.
+Flipping the banner on Vercel is "edit env var → redeploy", the same model as `SHOW_DRAFTS`. The banner `id` is namespaced per slug automatically (`hackathon-<slug>`), so prior dismissals reset cleanly between events.
+
+To stand up a new event: copy an existing file in `src/pages/hackathon/` to a new slug, edit its data (or write a custom layout), then set `HACKATHON_EVENT_SLUG=<new-slug>` and `HACKATHON_BANNER_ENABLED=true` on Vercel.
 
 ### Site URL Resolution
